Improve frontend resilience accessibility and performance
