+++ frontend/.gitignore
@@ -1,4 +1,5 @@
 node_modules
 build
+public/dist
 npm-debug.log*
 .env
