+++ frontend/public/website.html
@@ -32,7 +32,7 @@
         </div>
     </header>
 
-    <main class="landing-main">
+    <main class="landing-main" id="mainContent">
         <section class="landing-hero" data-animate>
             <div id="hero-section" class="hero-shell loading">Loading featured content...</div>
             <div class="hero-orb hero-orb--purple" aria-hidden="true"></div>
@@ -60,7 +60,7 @@
         <p>MovieNightPlanner · Plan better movie nights together.</p>
     </footer>
 
-    <div id="movieModal" class="modal" style="display:none;">
+    <div id="movieModal" class="modal" style="display:none;" role="dialog" aria-modal="true" aria-hidden="true" aria-labelledby="modalTitle">
         <div class="modal-content">
             <button class="modal-close close" type="button" aria-label="Close movie details" onclick="closeHomeModal()">x</button>
             <div class="modal-header" id="modalHeader">
@@ -80,5 +80,6 @@
     </div>
 
     <script src="app.js"></script>
+    <script src="js/homepage.js"></script>
 </body>
 </html>
